Refactor JS sources and build scripts
