+++ .jshintignore
@@ -1,5 +1,6 @@
 gulpfile.js
 **/*-min.js
+plugin/build/
 wordpress/
 node_modules/
 vendor/
