+++ package.json
@@ -33,18 +33,14 @@
 		"@wordpress/scripts": "14.0.1",
 		"babel-jest": "^26.0.1",
 		"babel-loader": "^8.1.0",
-		"css-loader": "^5.2.0",
 		"gulp": "^4.0.2",
 		"gulp-cli": "^2.3.0",
-		"gulp-minify": "3.1.0",
 		"gulp-wp-pot": "^2.5.0",
 		"ignore-emit-webpack-plugin": "^2.0.6",
 		"jest": "^26.0.1",
 		"jest-environment-jsdom-sixteen": "^1.0.3",
 		"jshint": "^2.12.0",
-		"mini-css-extract-plugin": "^1.4.0",
-		"node-sass": "^4.0.0",
-		"sass-loader": "^8.0.2",
+		"sass": "^1.53.0",
 		"terser-webpack-plugin": "^3.1.0",
 		"webpack": "^4.46.0",
 		"webpack-bundle-analyzer": "^4.2.0",
