+++ src/js/include/render.js
@@ -1,0 +1,3 @@
+export function renderResults(itemsToScan, activeOnly) {
+  // Todo in ENG-4
+}
